添加边缘检测

+++ 毕设/程序/run.m
@@ -12,5 +12,5 @@ if length(dot) >= 4
     img = correct_img(img, dot);
 
     figure;
-    imshow(uint8(img));
+    imshow(img);
 end
